refactor(view): extract widget_bridge_input.cpp (A1 first cut)
Pure structural extraction — no behavior change. `keycode_to_w3c_key`
(KeyCode → W3C UIEvent.key string) and `WidgetBridge::forward_key_event`
(101 lines total) move out of the 9.5k-line widget_bridge.cpp into a
focused widget_bridge_input.cpp.

Crucially this gives the input TU its own compat-surface mapping in
tools/scripts/compat_path_map.json — bound to `react` instead of the
`*` wildcard the legacy TU carries. Edits to focus-guard / key-event
forwarding no longer demand bypass trailers for canvas2d / css / html /
yoga / etc. (which was the #2120 pain — 7 bypass trailers for a 12-line
focus-guard change).

This is the *first cut* of A1. The remaining ~9.4k lines of
widget_bridge.cpp are dominated by a single 8.2k-line `register_api()`
body which dispatches widget/canvas/GPU/runtime registrations to the
JS engine. Splitting that safely requires a per-domain
`register_*_api()` method seam — a larger structural change that is
tracked in planning/2026-05-16-refactor-ops-high-leverage.md as the
"A1 follow-up" item.

Validation:
* widget_bridge.cpp shrinks from 9497 → 9396 lines (-101).
* widget_bridge_input.cpp = 130 lines.
* CMakeLists.txt adds the new source.
* compat_path_map.json maps the new file to a single `react` prefix.
* No semantic change to forward_key_event — the nested-if focus-guard
  is preserved verbatim.

Part of the refactor-ops batch.
