+++ tools/scripts/compat_path_map.json
@@ -8,6 +8,11 @@
       {"kind": "doc", "path": "docs/reference/compat/{prefix}.md"},
       {"kind": "test", "glob": "test/test_widget_bridge*.cpp"}
     ],
+    "core/view/src/widget_bridge_input.cpp": [
+      {"kind": "compat-json", "prefix": "react"},
+      {"kind": "doc", "path": "docs/reference/compat/react.md"},
+      {"kind": "test", "glob": "test/test_widget_bridge*.cpp"}
+    ],
     "core/view/include/pulp/view/widget_bridge.hpp": [
       {"kind": "compat-json", "prefix": "*"},
       {"kind": "doc", "path": "docs/reference/compat/{prefix}.md"},
